moving fake server to internal
