+++ managedkafka/go.mod
@@ -3,11 +3,9 @@ module github.com/GoogleCloudPlatform/golang-samples/managedkafka
 go 1.22.3
 
 require (
-	cloud.google.com/go/longrunning v0.5.7
 	cloud.google.com/go/managedkafka v0.0.0-00010101000000-000000000000
 	github.com/GoogleCloudPlatform/golang-samples v0.0.0-20240529190426-bba205b73ac2
 	google.golang.org/api v0.182.0
-	google.golang.org/grpc v1.64.0
 	google.golang.org/protobuf v1.34.1
 )
 
@@ -17,6 +15,7 @@ require (
 	cloud.google.com/go/auth/oauth2adapt v0.2.2 // indirect
 	cloud.google.com/go/compute/metadata v0.3.0 // indirect
 	cloud.google.com/go/iam v1.1.8 // indirect
+	cloud.google.com/go/longrunning v0.5.7 // indirect
 	cloud.google.com/go/storage v1.40.0 // indirect
 	github.com/felixge/httpsnoop v1.0.4 // indirect
 	github.com/go-logr/logr v1.4.1 // indirect
@@ -43,6 +42,7 @@ require (
 	google.golang.org/genproto v0.0.0-20240521202816-d264139d666e // indirect
 	google.golang.org/genproto/googleapis/api v0.0.0-20240528184218-531527333157 // indirect
 	google.golang.org/genproto/googleapis/rpc v0.0.0-20240528184218-531527333157 // indirect
+	google.golang.org/grpc v1.64.0 // indirect
 )
 
 replace cloud.google.com/go/managedkafka => ../internal/managedkafka/
